Create README.md
This README file was made on 2024-03-05 by Lia Bozza.
Title of Submission: Bozza Week 6 Take Home
This is the first R markdown file format that I have submitted for this class. The purpose of 
this assignment was to practice benchmarking & writing code, along with looking at ways to identify and replace characters in a matrix. Examples in the given document are vectorized. 
Instrument: RStudio4
Answer to final question: total "T"s in sequencing_matrix is 549,158. This code took 41.57172 seconds to run.